# openapi.generated2.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /api/v1/questionnaires/{id}:
    get:
      tags:
      - Questionnaires
      - Import
      - Export
      - Questionnaires
      summary: Get questionnaire metadata and screens index (no questions)
      operationId: getQuestionnaire
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/questionnaires/import:
    post:
      tags:
      - Questionnaires
      - Import
      - Export
      - Import
      summary: Import questionnaire CSV (v1.0)
      operationId: importQuestionnaireCsv
      requestBody:
        content:
          text/csv:
            schema:
              type: string
              format: binary
              title: Csv Export
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/questionnaires/{id}/export:
    get:
      tags:
      - Questionnaires
      - Import
      - Export
      - Export
      summary: Export questionnaire CSV (v1.0)
      operationId: exportQuestionnaireCsv
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError